Add 'skipFirstWait' argument to breakpoint decoration test helper
There are some test cases that navigate to  a source UI that restores
existing breakpoints. The "decoration update" event is fired before
the frame itself is available to the test, making it impossible to
instrument it safely without racing events.

This test adds a boolean  argument to an existing test helper. When set
we check breakpoint decorations immediatley without waiting for an
"decoratoin update" event. The decorations are still checked against
the expectation and a "wait" is queued up should not all expected
decorations be present.

R=mathias@chromium.org

Change-Id: I58e64e12557ac102b8c6e67152e18e15882277cd