Localization presubmit check to pass in one file-list instead of all files
In presubmit.py, _CheckDevtoolsLocalization() gets all the paths of affected
files and pass those in as arguments when running run_localization_check.py.
When a lot of files need to be scanned, it casus the issue where character
number exceed the limit of CreateProcess on Windows.

This PR changes the _CheckDevtoolsLocalization() to create a temp file that contains the affected file paths, and have the validation step read the
affected file paths from the temp file.


Bug: 941561
Change-Id: I698e5064287eee8a8a3aca0682e3ab3521eb63d1