Prepare AuditsService for audits_worker change
The audits_worker will be a worker with {type: "module"}. To prepare for
that change, we first need to properly expose the AuditsService on the
global. The previous variable declaration was local in the module scope,
which would cause resolution issues when trying to load a worker as
module.

We first need to push this to the appspot server, after which the
subsequent audits_worker change can be made.

Bug: 1006759
Change-Id: I0acd8ddae6b27907ab1b7c35af826fe3c6080261