[sources] Display the state of workers in the Threads section
Listen for the new CDP event that updated the state of workers and other
isolates.

Display the status next to the isolate name in the Sources panel. We
also show 'paused' here when paused in the debugger, so override the
thread state with 'paused' when paused, then go back to the thread state
when the isolate is resumed.

Bug: chromium:1025490
Change-Id: If68904ea830611ab97049304df75504455efacc4