[e2e] Moves shared logic out of e2e
We plan to have multiple suites using the same machinery that currently
runs the e2e. This CL moves that machinery out to a shared folder and
remaps the e2e code to use everything in the new locations.

Future CLs will introduce additional suites for capturing traces,
screenshots, and other artifact.

Change-Id: I452424d8f65fb03897f5a59606c592be996e91ab