Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> > Patch Set 3: -Code-Review
> > 
> > I've determined the issue is when NVDA is running, and that the issue persists with or without your patch.
> > 
> > @Amanda -- before we +1 this, can you share any tips about what might fix this particular issue?
> 
> @Rob -- Brian created a doc that proposed 4 different solutions (although I think one of them was invalidated). You could ask him about it because he's much more familiar with the bug than I am

Although I'm not familiar with this particular use of Dialog, I've worked on a related change with Dialogs before to modify the behavior of how it intercepts tab focus. Those changes are here.

https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1947173

For that bug, we ran into a similar issue where Settings Dialog was getting dismissed because actions were intercepted by events triggered outside the DOM. The click-outside-to-dismiss behavior can be disabled as a configuration option for the Dialog (see the CL).

Patch-set: 4
Reviewer: Gerrit User 1363999 <1363999@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>