Mismatched SASS source maps cause Performance recording to fail
A source file (.scss) is referenced, ostensibly by a compiled file,
and the .scss file is not available (it returns a 404). The
existing code casts-away a possible-null and then calls into
another function which cannot accept a null.

This change adds a null-check and skips removing the file from the
project if the source-mapped file isn't present.

Bug: 1049155
Change-Id: I9f0666b10c6c93eec095a53ee0f948a418752ef1