Add e2etest for Last-Modified parsing in network tab
It uses a `.rawresponse` file that forcibly sets the `Last-Modified`
header. We then open the network panel, add the `Last-Modified` column
to the data-grid and then introspect its values.

It also updates the contextmenu to set a proper `aria-label` on context
submenu's, which were missing. We use this label in the selectors of the
end-to-end test. This means it does not rely on DOM structure, but ARIA
structure.

This replaces
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/http/tests/devtools/startup/resources/cached-resource-metadata.html;drc=77578ccb4082ae20a9326d9e673225f1189ebb63?originalUrl=https:%2F%2Fcs.chromium.org%2F
as an e2etest.

Bug: 1047671
Change-Id: I89efc145c915f1a083dbcfb3765674122d9748fe