High contrast for Network tools panel
Within the network log, Headers, Initiator and Cookies,
I fixed the selected, hover and focus states. I opt-out on the color
bars in Timing to make sure we can see the bars.
For the timeline area, the indicator is set to Highlight and the
resizers will show Highlight on hover.
Also fixed the filter, where the one selected will be Highlighted.
Waterfall column was tricky, I did not update that column in this
pass.

Before: https://imgur.com/a/6R2vFFT
After: https://imgur.com/a/pfuOmSr

Bug: 1048378
Change-Id: I4acbea71b11da0f142fa7678762d97bd1e9289bf