DevTools: Fix forwarding keyboard shortcuts from target page
This was regressed by crrev.com/c/devtools/devtools-frontend/+/1942290
which removed an unused variable in response to a linting rule. That
variable referenced a call to construct a ForwardedInputEventHandler
instance. That constructor has a side-effect of registering event
listeners with the host and still needs to be run.

Restoring the call to the constructor fixes the issue.

Bug: 1049910
Change-Id: I6338164fcbbce6cae688d90c6d11cd76ed719f1e