[e2e] Add builder for experimenting with new features in the devtools recipe
While adding new steps to the existing recipe we run into trouble if the new step is not stable and fails. We do not want failures in CI or CQ. I propose we have this separate experimental builder on which we first try our new steps.

TBR: tmrts@chromium.org

Bug: chromium:1050027
Change-Id: Ie4e0e6aba858c6bd6700a23b55e2a8a9fced4ed1