Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1: Code-Review+1
> 
> (1 comment)
> 
> Consider my feedback "approved with suggestions".

I agree with the suggestion in general. I debated doing something like this, but decided not to given this regression has already shipped to stable in Chrome 80. In this case I think we should keep the change as minimal and close to the original code as possible in case we want to pull it into servicing.

Patch-set: 1