Update patch set 10
Patch Set 10:

Tabbing form the database on the tree to the left to the console selects the last evaluated statement instead of going directly to the new line. This prevents me from typing a new query, I need to press tab again.  I would expect the focus to be set on a new line ready to receive keyboard input

Patch-set: 10
CC: Gerrit User 1341877 <1341877@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>