Update patch set 1
Patch Set 1:

(2 comments)

The code changes LGTM.

I’ll defer to others to answer the question of whether or not we want to encourage more Server Timing usage in this manner. I worry slightly that a DevTools message like this would be interpreted as “Server Timing is a best practice now, use it OR ELSE!!1”.

For example, the spec itself includes this recommendation:

```
To minimize the HTTP overhead the provided names and descriptions should be kept as short as possible - e.g. use abbreviations and omit optional values where possible.
```

Thanks for kicking off this discussion, Jan!

Patch-set: 1
CC: Gerrit User 1214214 <1214214@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>