Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Tabbing form the database on the tree to the left to the console selects the last evaluated statement instead of going directly to the new line. This prevents me from typing a new query, I need to press tab again.  I would expect the focus to be set on a new line ready to receive keyboard input

I see, understood your concern. So like in the regular console, here too shift-tab selects the last prompt output, but typing takes the focus back to the prompt, which does not happen in the websql console. Possibly it is the code-mirror that is helping, but like discussed offline, seems more like a eature, not so much helping accessibility. So opened a CRbug for that - https://bugs.chromium.org/p/chromium/issues/detail?id=1050112, to be fixed separately.

Patch-set: 10