Update patch set 26
Patch Set 26:

This is a very large CL and it is difficult to understand why certain changes are made. Moreover, the screenshot appears to differ in what functionality it shows.

Could you split this up in self-contained CLs that target 1 particular component and provide screenshot before/afters for these components alone?

Patch-set: 26
CC: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>