Update patch set 3
Patch Set 3: Commit-Queue+2

> Patch Set 3: Code-Review+1
> 
> Less code, less confusion, I like it! Will you delete the prototype hijacking in a follow-up CL?

This is only a subset of the usages of {valuesArray} mainly "for of" loops. The other use-sites need to be either converted to Array.from(<var>.values()) or handled differently.

Moreover we also need to check web tests for usage. What makes this a bit cumbersome is that {MultiMap} only implements {valuesArray}, but not {values}. Meaning each use-site needs be checked if a Map/Set or a MultiMap is used.

Patch-set: 3
Reviewer: Gerrit User 1323977 <1323977@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2