Update patch set 26
Patch Set 26:

> Patch Set 26:
> 
> This is a very large CL and it is difficult to understand why certain changes are made. Moreover, the screenshot appears to differ in what functionality it shows.
> 
> Could you split this up in self-contained CLs that target 1 particular component and provide screenshot before/afters for these components alone?

Tim, yes, I was afraid that this might be getting bigger. The majority of the work is for Sources tool. Updates to others tools happened, due to treeoutline.css being updated. I will take a pass to break up this CL. Just a thought, I feel maybe treeoutline could be a CL in itself, since it touches practically most tools, in some way.

Patch-set: 26