Add high contrast styles for Memory tools panel.
The sidebar got an update to make sure the elements has the correct
color values on hover and selected. The radio button's background is
updated to be Highlight color when checked.
The JavaScript VM instance also go a pass to make sure hover, selected
and focus states are visible.

Before: https://imgur.com/a/AI37KnN
After: https://imgur.com/a/3kTtlAb

Bug: 1048378
Change-Id: I1772116f313b43edfe5bf3f649c081bf2175b963