Add test to verify multiple workers are shown as different entrypoints
The test inspects a page that spawns 2 workers. Each worker is in a
different "domain" show adjacent to the source tree of the loaded page.
It also updates the source tree aria-labels to properly show the full
url path that leads up to that particular node.

This replaces
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/http/tests/devtools/startup/dedicated-workers-list.js;drc=77578ccb4082ae20a9326d9e673225f1189ebb63?originalUrl=https:%2F%2Fcs.chromium.org%2F
as an e2etest.

Bug: 1044632
Change-Id: I46db1af836dbb6f20fefc0d4c1913cd82d3cb2ce