Settings Discoverability and Telemetry
This CL adjusts entrypoints for Settings.

1) Adds a Settings Gear icon to the top-level toolbar
2) Moves Settings ... menu entry to More tools submenu

https://i.imgur.com/9WS6EGy.png

This CL also includes telemetry for Settings,
so we can better evaluate how users are finding Settings,
and which Settings tabs are most popular.

1) Record a User Action for opening Settings
   - Reveal Settings via Gear icon
   - Reveal Settings via Top Menu -> More tools -> Settings
                  OR via Top Menu -> Shortcuts
   - Reveal Settings via Command menu

Remark: The keyboard shortcut F1 / Shift+? to reveal Settings
does not record an action here; keyboard shortcut telemetry for the
settings.show action is already taken care of.

2) Report Panel Open action for each Settings tab
   - On reveal: should report exactly the panel opened once
   - On switch tabs: should report exactly the panel switched to once

The corresponding Chrome CL to add the new enums is here.
https://chromium-review.googlesource.com/c/chromium/src/+/2055359

In order to have the Settings screen record Panel Open actions
reliably, a new TabInvoked event was added to TabbedPane which is
triggered by all tab selections, both programmatic and user invoked.

The existing TabSelected event only triggers on tab changes.
Using TabSelected prevent the Settings screen from reporting panel open
events when closed and re-opened to the same tab.

One criteria for determining whether to report a panel open is that
the tab is invoked by a user (userGesture) and not programmatically.
For user invokable items in context menu and the Command menu,
this userGesture argument is now set to `true` by default,
as both must have been invoked by a user.


Bug: 1050855
Change-Id: Id2e6e5f995bf2e6d419a3144eeec3e6af2bf208b