Settings Discoverability and Telemetry
This CL adjusts entrypoints for Settings.

1) Adds a Settings Gear icon to the top-level toolbar
2) Moves Settings ... menu entry to More tools submenu

https://i.imgur.com/9WS6EGy.png

This CL also includes telemetry for Settings,
so we can better evaluate how users are finding Settings,
and which Settings tabs are most popular.

1) Record a User Action for opening Settings
   - Open Settings via Gear icon
   - Open Settings via More tools
   - Open Settings via Command menu

2) Record Panel Open action for each Settings tab

As this is the first time we have recorded telemetry which
distinguishes between the entry points on the same action,
this CL aims to strike a balance between adding additional
infrastructure to base classes (Toolbar, CommandMenu) and
injecting special telemetry for this specific case
(MainImpl, CommandMenu).

Bug: 1050855
Change-Id: Id2e6e5f995bf2e6d419a3144eeec3e6af2bf208b