Treeoutline - Centralize common styles and update selected color.
Tree outline does not have a hover state. Hover state is done inside
the tools that has a hover set. Tree outline does have a common styles
to set hover on foreground text and icons.
Tree outline does have focus, selection and selected. Focus and
selection has the 'Highlight' color treatment. The selected state,
is set to 'ButtonText' color. This will give the impression that
the button is selected, but not focused.

Change-Id: I574e5fc921045a836546c224ac31e38d8eb3f8ff