Add high contrast fix in Performance tools panel.
For the pie chart and its labels, it is opt-out of forced colors
on the color swatch, so we are able to distinguish the difference of
swatches, instead of seeing empty boxes. Timeline status dialog
got a border treatment and fixed text-buttons to use the correct
color values for hover and focus.
Hover, select, and focus on grid view, for Bottom-Up, Call Tree
and Event Log, is updated to correct color value.

update:
create css variable for box shadow, use border-color when changing colors on borders

Before: https://imgur.com/a/0sD79Qy
After: https://imgur.com/a/vxsIo2m

Bug: 1048378
Change-Id: If9d89702a6eb427ab8e6d240c087fe70dc85d867