Get a RuntimeModel test working with Sinon.
This adds a test that mocks enough stuff out (and there's a lot) to be
able to test `RuntimeModel` and use Sinon to check on an
`invoke_evaluate` call.

This CL is not to be merged as is but more to trigger a discussion on
next steps and what we do from here to really get a good mocking story
in our unit tests for the future.

Change-Id: I09a65c3d8b853b6291dadf407e432855d9df38e6

DO_NOT_SUBMIT