Issue 1050999: Properties Tab
This bug was introduced with a previous change. What happened was,
the code previously was a "fake" tree outline, in that it stacked
a bunch of ObjectPropertiesSection objects on top of each other,
but there wasn't a root tree. In the change, I added in a new tree,
but neglected to note that the operation that cleared out the
virtual "tree" in this case deleted the DOM reference to the new
real tree that I added in. This seemed to only occur when a refresh
occurred, not when changing nodes in the Elements tree.

The correction here is to have the actual tree clear its children
out rather than the container.

Bug: 1050999
Change-Id: Icff91d1a76f3d8fe71d78d1c22ad614617a25af3