Add http/tests/devtools/user-metrics.js webtest.
This test is originally in Chromium, but it is testing constants defined
in the devtools frontend code. The test should therefore be directly in
the frontend repo.

With this (plus a followup CL to remove the test from Chromium), this CL
can land without conflicts: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2044605

Change-Id: I4cca763b0e768a73d736fefd8a141357fbe11ab5