Update patch set 3
Patch Set 3: Code-Review+1

LGTM, but once this lands I will need to migrate this code to ES modules to keep it in line with the rest of the codebase. I'll cc you in on that CL so you know, but I mention it because it could cause merge conflicts if you didn't know.

Patch-set: 3
Reviewer: Gerrit User 1189020 <1189020@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1