Don't scroll when focusing tree outline
The algorithm for focus() calls ScrollIntoView on the element that's
focused. Prior to r718454, ScrollIntoView used a magic number to avoid
scrolling if an element was a certain number of pixels on screen
already. r718454 removed this magic number to improve interop but this
means that focus calls in DevTools' tree outline may now cause
horizontal scrolling in an otherwise unscrollable widget.

This CL simply makes explicit that we shouldn't scrollIntoView on focus
by passing the preventScroll option to it.

Bug: 1036817
Change-Id: I4026b88109a2b3e76869dfae5921e1798386cfb1