ux(devtools): scroll console scopes if more exist than can be seen
The scope menu is restricted to a specific height. However if a user
was to have more than would fit in that, they were unable to reach them.

This patch applies automatic overflow on the y axis to the menu.
If the amount of scopes fits in that height, they will have no scrollbar.
A scrollbar will appear when more than can fit within the menu are present.

This allows users to get access to all scopes available regardless of the menu height.

Bug: 1038461

Change-Id: Ia36b00d0b991a9614c1c7a2a2cf6e95540ca75e1