Update patch set 18
Patch Set 18:

Let's discuss in person what the details are of this change. I am not against it, but it would be good to define where what should live, to reduce confusion. E.g. we call `target.goto` in the test itself, but do other navigation in the objects.

Thanks for looking into this!

Patch-set: 18
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>