Fix focusing stop button after timeline profiler initialization
This was regressed by crrev.com/c/1971936 which delayed enabling the
stop button until the profiler was initialized. The call to focus the
stop button still occured when showing the status pane, but failed to
work as the button was currently disabled.

Fixed by moving the call focusing the stop button to occur immediately
after the button has been enabled.

Bug: 963183
Change-Id: I59a1a6ac5f02e7c57ace59111f90458f757f7fd0