High contrast pass to dropdowns, buttons and links
There are multiple changes in this patch. For the Styles panel,
the color on the the web property when focused, now changes to
HighlightText. The link is now LinkText color on rest and hover state,
but changes to HighlightText when focused.
The toolbar with drop-down, the history drop-down and search box
now have borders. The disabled text button is fixed, with the correct
background, foreground and border color. The toolbar button text has
the correct color on hover and focus.
Made another pass to remove !important

Before: https://imgur.com/a/PJRm3K0
After: https://imgur.com/a/kFUXCFp

Bug: 1048378
Change-Id: Ie33778a0129203b98c5426f2a4a58fc5ee34a568