Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+1
> 
> LGTM
> 
> As a side note, any thoughts on how we might want to add tests for high contrast support in DevTools? We have e2e tests now [1], and Puppeteer supports emulating media features [2]. Paul has recently been working on screenshot testing support. Is anything missing for you, Kalon?
> 
> [1]: https://source.chromium.org/chromium/chromium/src/+/master:third_party/devtools-frontend/src/test/e2e/test-list.ts
> [2]: https://github.com/puppeteer/puppeteer/blob/master/docs/api.md#pageemulatemediafeaturesfeatures

Out of one of our syncs we were given a head's up that screenshot testing was in the works. We would love to leverage that for high contrast and general theming validation, so it's good to hear progress is being made.

I skimmed the test folder and I think all the pieces are in place [1] for us to write some tests. I'll start with a simple test for the Elements panel. While I experiment, I'll chat with other Edge folks to come up with a list of scenarios we want to test and run it by y'all.

[1]: https://source.chromium.org/chromium/chromium/src/+/master:third_party/devtools-frontend/src/test/shared/screenshot.ts;l=29-46

Patch-set: 5
Reviewer: Gerrit User 1341869 <1341869@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>