Update patch set 2
Patch Set 2:

I'm a big +1 to adding `_test` to our test files, makes it easier to navigate/find/differentiate/etc.

For the tests that depend on `platform/utilities`, is the fix to import that file in those tests? (I'm not saying that's ideal), or is there more to it?

Patch-set: 2
Reviewer: Gerrit User 1396442 <1396442@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>