Rewrite ElementsBreadcrumbs as a TypeScript custom element
This is the first attempt at building a piece of UI in the devtools with
TypeScript and custom elements.

This CL also explores how we can interface between the old code and the
new, using events as required.

Much of the code here (particularly in the `_bridge.js` file) should be
auto-generatable, but one-step at a time :)

DISABLE_THIRD_PARTY_CHECK=TypeScript build fixes

Change-Id: I122fab727a3fa9e45d34df2c5690032685504847