Fix keyboard focus loss when using JS profiler filter buttons
The focus, exclude, and reset buttons for filtering the functions in a
profile each become disabled once used, causing focus to be lost.

Fixed by focusing one of the remaining, enabled toolbar items. For most
cases this is the reset button. When the reset button itself is used
focus now goes to the combo box used to sort functions.

After changing focus data-keyboard-focus was still not being set. This
was due to focus changing in response to a click handler triggered by
the keyboard event, a scenario likely to exist in other tools as well.

Fixed by updating the helper utility for tracking keyboard focus to
account for keyup events in addition to keydown events. Since keyup
fires at the end of the keyboard interaction, it extends the scope of
events covered to include the triggered click event.

Before: https://imgur.com/a/yxyv0i4
After: https://imgur.com/a/cHpwVvs

Bug: 1054107
Change-Id: Ic3e323b06c534cabfbde00f147baa904db6ccefa