Fix disappearing focus in Performance and Coverage
This CL fixes two similar issues where focus disappears or moves to
invisible objects. The first is that Performance panel timeline
currently keeps tabstops while blank, creating the appearance of focus
being lost after tabbing past the "Learn more" link. Hiding the
flamechart widget when it has nothing to show avoids focus moving to it.

In the coverage drawer, focus is lost completely when the user presses
the reload and record button. To avoid this, I've made the coverage
drawer check for focus when a recording begins and focus the datagrid if
the drawer has focus.

Video of performance working properly: https://i.imgur.com/OQKmeh9.mp4

Bug: 1055375
Change-Id: I553d48cad29611ae1828bc423f0d6ba7e7a6383e