Add telemetry for keyboard shortcuts
This CL adds telemetry for keyboard shortcut usage, which will help us
evaluate the success of the upcoming custom keyboard shortcut and
settings discoverability experiments. Discrete actions are added for
each of what we expect to be the most-used shortcuts in the most-used
tools (e.g. debugger.step), and other shortcuts are recorded in an
OtherShortcut bucket.

Change-Id: Ia7bd751912fd39a09d330707e33fcc1ab13eebcf