Add toolbar showing count of warnings to issues panel
Add a toolbar to the issues pane. Its only content for now is an icon
with the number of detected issues next to it. More content will be
added later. Explicitly displaying a count of 0 fixes the confusion
caused by the issues pane being completely empty when there are no
issues to display.

Fixed: chromium:1053466
Change-Id: I8a325f651d15f76960841b9b69a0693eb53fd63f