Run TypeScript without building mode
We have been seeing spurious CQ runs that complain that Ninja leaves dirty
files behind. Re-reading the documentation [1] leads us to believe that
the usage of the `--build` flag is incorrect. We should led Ninja perform
these deps check, rather than TSC.

Changing the flag to `-p` (which is normal compilation) appears to work
fine. We will need to do additional profiling on the effect of this
change, but we are trying to figure out if this resolves the CQ failures
first, as they are blocking CL submissions.

[1]: https://www.typescriptlang.org/docs/handbook/project-references.html#caveats-for-project-references

Bug: 1011811
Change-Id: I473d610d042af1e0aba22f3d2a260eac4ceb9369