Switch the Media tab to ES Modules
crrev.com/c/2057206 broke the media tab by adding an import statement to a
non-module file. I figured that this was as good a catalyst as any to
switch to ES Modules. They seem to be the direction the frontend UI is
going and there doesn't seem to be any regression testing to prevent
breaking changes like this as things keep moving that way, so might as
well get out in front of it.

Bug:1054925
Bug:1006759
Change-Id: I99785423335d234a3d87743e08768f5f8cf56bce