Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

-1 only to make sure that you don't accidentally miss it based on Jacks review.

In general, we shouldn't be making the types more generous solely for pleasing the TypeScript compiler, as it is hiding errors. At that point, we are not gaining any value from our type checkers and are making it worse.

I am happy to take a stab at this if you are stuck, as it seems to be a bit more involved that I initially estimated.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1