Update patch set 4
Patch Set 4:

(1 comment)

Thanks for looping me in. Please revert the changes to CompilerScriptMapping. The asyncification work is done top-down, so we can split up the work. This means that we are awaiting methods that are not yet async.

Patch-set: 4
Reviewer: Gerrit User 1323977 <1323977@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>