Update patch set 4
Patch Set 4:

Quick question Michael, I am noticing that before this change, the behavior was that once you use keyboard to click the button, in case you want to press stop after reload, all you have to do is tab around one or two times, and focus comes back to the button. With this change, however, focus goes to the table, so now you have tab all the way back to the button to toggle the reload button. Is there some way to being the focus back to the button itself instead of the table of contents?

Patch-set: 4
CC: Gerrit User 1334577 <1334577@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>