Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This library doesn't have a LICENSE file, making it hard for us to be sure it's okay for use. It does declare itself as MIT in the package.json file, but we should file a bug in the GitHub repo asking if the author minds adding a dedicated LICENSE.

So I will use my personal account to do that and I prepared the following:

Title: Dedicated LICENSE file is missing

Description:
I noticed that you declared in the packacge.json file that it is MIT, however I noticed that you do not have a dedicated LICENSE file. We would like to use your work in another open source project and we need to make sure that it's okay for us to do so.

can you let me know if it is okay to post it like this?

Patch-set: 3